work in progress

+++ build-overlays.sh
@@ -6,13 +6,15 @@ dir=`dirname $0`/src/overlays
 cd $dir
 dir=`readlink -f .`
 
+mkdir -p ../../build/overlays
+
 if [ -z "$1" ]; then
 	echo "Building all overlays..."
 	for i in `find . -maxdepth 1 -mindepth 1 -type d`; do
 		# remove './' from begining
 		x=${i#./}
 		cd $x
-		rm ../../../build/overlays/$x.tar.gz -f
+		test -e ../../../build/overlays/$x.tar.gz && rm ../../../build/overlays/$x.tar.gz -f
 		tar cvzf ../../../build/overlays/$x.tar.gz .
 		cd $dir
 	done
@@ -20,7 +22,7 @@ else
 	echo "Building overlay $x"
 	x=$1
 	cd $x
-	rm ../../../build/overlays/$x.tar.gz -f
+	test -e ../../../build/overlays/$x.tar.gz && rm ../../../build/overlays/$x.tar.gz -f
 	tar cvzf ../../../build/overlays/$x.tar.gz .
 	cd $dir
 fi
